fix - create account

+++ src/main/java/com/bank/bank/dto/ContaRequestDTO.java
@@ -2,4 +2,4 @@ package com.bank.bank.dto;
 
 
 
-public record ContaRequestDTO (Integer saldo, String numero, Long idCliente) {}
+public record ContaRequestDTO (Double saldo, String numero, Long idCliente) {}
